+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * chore: migrate from deprecated api
 * fix: fixed issue with type check in prefer-match-file-name
+* fix: returns 0 exit code while found unused files or l10n in cli mode
 
 ## 4.8.0